refactor(persistence): split snapshot.rs into submodules (TD-025)
Behavior-preserving decomposition of the 1218-LOC monolithic
snapshot.rs into a module hub with three focused submodules:
- snapshot/types.rs   — serializable data structs + serde helpers
- snapshot/convert.rs — NpcSnapshot::from_npc / into_npc
- snapshot/restore.rs — GameSnapshot::capture / restore + privates
- snapshot/tests.rs   — all tests moved alongside the logic

Public API re-exported unchanged from snapshot/mod.rs. Part of #1200.
